requirements.txt
google-cloud-bigquery==3.30.0
google-cloud-aiplatform==1.82.0
google-cloud-storage==2.19.0
google-auth==2.38.0
streamlit==1.43.1
pandas==2.2.3
plotly==5.22.0
numpy==2.2.3
requests==2.32.3
protobuf==5.29.3
db-dtypes==1.4.2